# spec/swagger.yaml
definitions:
...
  github_com_Kxiandaoyan_Memoh-v2_internal_mcp.Connection:
    properties:
      bot_id:
        type: string
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  handlers.BatchDeleteRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  mcp.ExportResponse:
    properties:
      mcpServers:
        additionalProperties:
          $ref: '#/definitions/mcp.MCPServerEntry'
        type: object
    type: object
  mcp.ImportRequest:
    properties:
      mcpServers:
        additionalProperties:
          $ref: '#/definitions/mcp.MCPServerEntry'
        type: object
    type: object
  mcp.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_Kxiandaoyan_Memoh-v2_internal_mcp.Connection'
        type: array
    type: object
  mcp.MCPServerEntry:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      cwd:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      headers:
        additionalProperties:
          type: string
        type: object
      transport:
        type: string
      url:
        type: string
    type: object
  mcp.UpsertRequest:
    properties:
      args:
        items:
          type: string
        type: array
      command:
        type: string
      cwd:
        type: string
      env:
        additionalProperties:
          type: string
        type: object
      headers:
        additionalProperties:
          type: string
        type: object
      is_active:
        type: boolean
      name:
        type: string
      transport:
        type: string
      url:
        type: string
    type: object
...
info:
  contact: {}
  title: Memoh API
  version: 1.0.0
paths:
...
  /bots/{bot_id}/mcp:
    get:
      description: List MCP connections for a bot
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mcp.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List MCP connections
      tags:
      - mcp
    post:
      description: Create a MCP connection for a bot
      parameters:
      - description: MCP payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/mcp.UpsertRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Kxiandaoyan_Memoh-v2_internal_mcp.Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Create MCP connection
      tags:
      - mcp
  /bots/{bot_id}/mcp-ops/batch-delete:
    post:
      description: Delete multiple MCP connections by IDs.
      parameters:
      - description: IDs to delete
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchDeleteRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Batch delete MCP connections
      tags:
      - mcp
...
  /bots/{bot_id}/mcp/{id}:
    delete:
      description: Delete a MCP connection by ID
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Delete MCP connection
      tags:
      - mcp
    get:
      description: Get a MCP connection by ID
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Kxiandaoyan_Memoh-v2_internal_mcp.Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get MCP connection
      tags:
      - mcp
    put:
      description: Update a MCP connection by ID
      parameters:
      - description: MCP ID
        in: path
        name: id
        required: true
        type: string
      - description: MCP payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/mcp.UpsertRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Kxiandaoyan_Memoh-v2_internal_mcp.Connection'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Update MCP connection
      tags:
      - mcp
  /bots/{bot_id}/mcp/export:
    get:
      description: Export all MCP connections for a bot in standard mcpServers format.
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mcp.ExportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Export MCP connections
      tags:
      - mcp
  /bots/{bot_id}/mcp/import:
    put:
      description: Batch import MCP connections from standard mcpServers format. Existing
        connections (matched by name) get config updated with is_active preserved.
        New connections are created as active.
      parameters:
      - description: mcpServers dict
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/mcp.ImportRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mcp.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Import MCP connections
      tags:
      - mcp
...
swagger: "2.0"